Fixed issue with player starting to play after unlocking
For some reason, AVAudioSessionInterruption (.began) is received,
followed almost immediately by AVAudioSessionInterruption (.ended)
when the application is opened after play was paused outside of the
application (e.g. in control center when device is locked).

So, don't set pausedDueToAudioInterruption if the player is
not playing when .began is received (so playing will not
restart when .ended is received).